✅: Split slow test from fast test class
