+++ selidor-projects/selidor-libraries/selidor-identifier/src/test/java/pw/itr0/selidor/identifier/codec/LowerCrockfordBase32Test.java
@@ -67,29 +67,6 @@ class LowerCrockfordBase32Test {
         });
   }
 
-  @Test
-  @Tag("slow")
-  void testRandomBytes() {
-    assertSoftly(
-        s -> {
-          final SecureRandom random = new SecureRandom();
-          for (int i = 0; i < 1000; i++) {
-            final byte[] r = new byte[16];
-            random.nextBytes(r);
-
-            // byte[] <-> byte[]
-            final byte[] bytes = LowerCrockfordBase32.encodeToBytes(r);
-            s.assertThat(bytes).hasSize(26);
-            s.assertThat(LowerCrockfordBase32.decode(bytes)).isEqualTo(r);
-
-            // byte[] <-> String
-            final String string = LowerCrockfordBase32.encode(r);
-            s.assertThat(string).hasSize(26);
-            s.assertThat(LowerCrockfordBase32.decode(string)).isEqualTo(r);
-          }
-        });
-  }
-
   @SuppressWarnings("SpellCheckingInspection")
   static Iterable<Arguments> base32EncodeLongToStringMapping() {
     return List.of(
